feat: create masked input
